+++ docs/release_notes.rst
@@ -30,6 +30,12 @@ RUS 12
 Expanded Data Coverage
 ^^^^^^^^^^^^^^^^^^^^^^
 
+EIA-860M
+~~~~~~~~
+
+* Updated EIA-860M with monthly data through January 2026. See :issue:`5042` and
+  :pr:`5044`.
+
 Documentation
 ^^^^^^^^^^^^^
 
@@ -38,6 +44,10 @@ Documentation
 
 New Data Tests & Validations
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+* Normalized RUS-7 and RUS-12 borrower ID's, names and state in
+  :ref:`core_rus7__entity_borrowers` and :ref:`core_rus12__entity_borrowers`.
+  See :issue:`5040` and PR :pr:`5056`.
 
 Bug Fixes & Data Cleaning
 ^^^^^^^^^^^^^^^^^^^^^^^^^